Fixes no issue in shared issue panel when reopening project.
If the linter reports the issue after layout editor render the model
the reported lint issues are miss because ModelLintIssueAnnotator
doesn't update the information without rendering.

Make ModelLintIssueAnnotator annotates the issues to NlModel when
background highlighter is triggered.

This CL add back the background highlighter which is removed by
http://ag/19340848. Now triggering background highlighting or complete
the rendering both update the issues.

Consider to use MergingRenderingQueue to control the updating events in
the following CLs.

Bug: N/A
Test: manually
Change-Id: I5368d594805465dcc8594c38428aec46779e4e47