Add snapshot coverage for KMP unresolved references
Modified the existing snapshot project nonStandardSourceSetDependencies
to cover the scenario when a KMP Android Library module depends on
KMP JVM module.
Rename the :network-kmm module on the project to :common
in order to resolve the KMP JVM module before the Android Library causing
the unresolved reference issue.
Update ideModels, projectViews and syncedProject comparisons .text with
the expected output for the changes.

Bug: 234309878
Test: snapshots updated and manual validation
Change-Id: If46b1756de5dd015a6dd2a682c24a656f5fb1de3