Filter device state by ScreenOrientation instead of name
In updateConfigurationScreenSize, we select the device State orientation
depending on the vertical and horizontal dimensions received. Since this
function was written before the device spec was introduced, we relied on
the fact that landscape states were always named "Landscape", while
portrait states were named "Portrait".

Since device states created from a device spec declaration do not comply
with the assumption above, which was actually never a contract, we need
to fix the filter to check for the orientation attribute when selecting
the state based on the orientation, otherwise we would throw a NPE.

Fixes: 241567480
Test: N/A. Manually verified.
Change-Id: Ib98c40a0023de6ce630d4ba7b0deda56a29c44ad