Throw exception when merging artifactsToModuleMaps with duplicate keys
In case IntelliJ by mistake adds the same artifact into kmp and
non-kmp maps respectively MPP_CONFIGURATION_ARTIFACTS and
CONFIGURATION_ARTIFACTS, and the value is different, then we are
going to throw an error in order to avoid override the values.

Something that we potentially should be considered in case they want
end up mixing artifacts is to combine the values, given that the
artifacts are mapping with a list of modules that can be combined.

Bug: 234309878
Test: already covered

Change-Id: I18962f9ae988fa47ec9a395c0d5b23230e0b5bbb