Add a generalized traversal for MultiPreview
This change adds a generlized way of traversing the annotations
graph of a given project, that will later be used to refactor the
current implementation of Compose MultiPreview, and could also be
used by other Preview tools to implement a MultiPreview logic,
reusing the traversal of the annotations graph that is being
added here.

Note that this CL only adds the AnnotationsGraph logic, but it is
not being used yet.

Bug: 228294269
Bug: 238749730
Test: AnnotationsGraphTest
Change-Id: I85d10f4773d535d6069fff5d32b968e098da1c12