[SDK Index] Add E2E for PSD integration
... similar to lint SdkIndexLintTest.

SDK Index initialization can happen when a build.gradle file is open in
the editor and when PSD is shown. We currently have a test to confirm
the initialization is done by lint when opening a build.gradle in the
editor. This change adds a similar tests but instead of opening the
editor it opens the PSD. Due to b/242358149, test is disabled on Mac.

Bug: 233789469
Test: SdkIndexPsdTest, SdkIndexLintTest
Change-Id: I392afca72aab2ebf21812a12baed181955e4f2b4