Replace ApkFacetCheckerInitializer with a service
ApkFacet is defined in android-core. In order to check whether a project
has ApkFacet from android-common, a separate interface (ApkFacetChecker)
was introduced in I86d0307617. But the existing implementation relies
on some initialization happening very early during IDE startup.

This CL simplifies this by using a project service whose implementation
resides in android-core, but the interface resides in android-common.
While it solves the problem, it does require android-common to implicitly
assume that someone will implement this service.

Bug: 200037669
Test: launch Studio, verify that apk facet verification still works
Change-Id: I09d9a9cd1d869ea93a8dd57bda82b59e3a2cb030