Update 'download sources' banner for new scenarios
This banner currently displays when the user has open a decompiled class file from the current compile SDK. The class file is always a stub, and this banner allows the user to download sources for the SDK instead.

This change extends functionality to also offer to download sources for a different SDK/API level, when the user is debugging on a device with a different API level. AndroidPositionManager is already detecting that scenario and generating a file to be displayed for end users (instead of nothing, or worse a mismatched file); it now will add metadata to that file that AttachAndroidSdkSourcesNotificationProvider can use to display itself.

I've largely left the existing decompiled class scenario for this banner alone, with two minor changes. First, strings have been updated and simplified. Second, the "Refresh" link has been removed, since its logic will be automatically run after sources are downloaded and its functionality is unnecessary. Both of these changes behind a studio flag.

Finally, refresh for source files coming from AndroidPositionManager is *not* implemented yet, and will come in a subsequent change.

Bug: 74128300
Test: Included
Change-Id: Ic0281c8083c07530c07d8eb062604a1050455554