Create copy of AndroidPositionManager
This change is creating a copy of
debuggers/src/com/android/tools/idea/debug/AndroidPositionManager.java.
The new file (AndroidPositionManagerOriginal.java) is an exact copy; git
likely will not show it as such since the original file still exists and
this is not a move operation.

The intent of this change is to allow both an "old" and a "new" version
of this file to exist in parallel as feature development is done on the
new version. The old version (AndroidPositionManagerOriginal.java) is
not expected to have any new changes, and will eventually be deleted
once the new version is ready.

The corresponding test file is being copied in a similar manner.

Change-Id: Ia5af4ded1c09ae95af4192d82f44bfa2cf22e203
Test: Test is copied as well as source
Bug: 74128300