Separate NotificationHyperlink and SyncMessageHyperlink
which are used in in different scenarios and one needs a sync issue
quick fix id while the other does not.

This is to allow further refactoring and simplification of the sync
issue reporting pipeline in order to make it better testable and support
multiple Gradle roots in an IDE project.

Note, we drop quick fix ids in some hyperlinks as they never reach
`GradleSyncMessages` and the value becomes unused.

Bug: 242440055
Test: existing
Change-Id: I39ef54d96079385f17894afe24bf9d29587a6c5e