Add support for multi-hyperlink handlers
`SyncMessages`s are constructed from a text message and a collection of
quick fixes represented as hyperlinks. However, in some cases the
message itself may need to contain hyperlinks and in this cases quick
need to handle already created hyperlinks.

Introduce `SynMessageNotificationHandler` to handle hyperlinks without
containing the hyperlink itself. Keep the `SyncMessageHyperlink`
compatible with the existing code to avoid cascading changes.

Bug: 242440055
Test: existing
Change-Id: Ie4f8e6656cccd5524fbad8b59a79d7c3d8152eda