Don't attempt to re-assign TODO toolwindow's shortcut to logcat
A long while ago we used to switch <mod>-6 from TODO to logcat.
This code seems to have bit rotted and is no longer functional:
the TODO toolwindow doesn't come with a default shortcut, <mod>-6
goes to problems, the action id to activate the logcat toolwindow
is invalid. This CL just gets rid of that whole code.

See b/212740693#comment3 and b/36356796 for prior issues around
this.

Bug: 243082216
Test: no shortcut for logcat toolwindow with or without this CL
Change-Id: I923aeb356c850cc6b338900320ff5bdbde560316