Restore the ic_launcher_round.xml
Once ic_launcher_round.xml was removed as part of the effort to
introduce the themed app icons.
But, removing it disabled rounded icons for devices < API level 26.

This CL restores the ic_launcher_round.xml with the monochrome attribute
so that themed app icons are effective even though ic_launcher_round.xml
is specified in the manifest.

Also removes the v33 variant of the ic_laucher.xml because adding a
property for v26 variant is enough for applying themed app icons.

Bug: 218931654
Test: Manually created a project and verified ic_launcher.xml and ic_launcher_round.xml existed.
Change-Id: I59de5aff7edbb9cfccc7e755cfcf021aedcfb06b