Memory reporting: load agent only before the usage
Not all the IDE sessions will provide memory usage reports: for some of them the server side flag will be disabled, some IDE sessions will be <30 min. Because of that we don't need to attach the helper JNI agent at the IDE startup, we can do it just before the report collecting traversal.

Bug: 235360502
Test: None
Change-Id: Ice332e16678e917b7ed85ae49ea3d2778a33a503