Update SourceCodePreview representations when smart/synced
In SourceCodePreview, we update the list of PreviewRepresentation that
are supported by the current file. That means we'll call the accept()
method of multiple PreviewRepresentationProvider, which often depends
on the project being smart and synced to calculate its return value. For
example, ComposePreviewRepresentationProvider needs to read a property
(usesCompose) from AGP flags to determine if should accept the current
file.

Updating the representations before being smart and synced means we
can't properly determine if some PreviewRepresentationProvider are able
to represent the file, and this might have a negative impact on the UX,
for example hiding the split view controls in Compose files.

This change ensures we updaet the SourceCodePreview representations
when smart and synced.

Fixes: 240196627
Test: N/A. Manually verified.
Change-Id: I0fea977087031ba01c75c2c9e07474275095c0ae