[AUA] Extract and test a utility function
In preparation for its use elsewhere in the Upgrade Assistant.  We
need to give this utility some entry point that is public in order to
be able to write unit tests; make it clear that that entry point is
not for general consumption.

In the process, fix a couple of minor bugs in the utility uncovered by
writing tests.

Bug: n/a
Test: added, existing
Change-Id: I6c6612179ed7fe0da4a128d8b7349128ffe63fa4