Fix Kotlin version used by the new project wizard
Recently, JetBrains has started using a different Kotlin compiler
version for IDE analysis than for JPS compilation. That is, the
Kotlin compiler bundled in kotlin-plugin.jar is different from
the Kotlin compiler under plugins/Kotlin/kotlinc.

We were using the "IDE" Kotlin version when creating new Android
projects, which currently results in using version "1.7.20-dev-1059".
Instead we should be using the "JPS" version, which will in general
be a stable Kotlin version. This matches the behavior of the
new project wizard in IntelliJ as well.

Bug: 246007737
Test: several UI tests such as CreateBasicKotlinProjectTest
Change-Id: I18c1b7129fc42481e1f1f19c0b19f99aa9535ca2