Remove unnecessary encoding for string resource
This panel is used in a dialog for creating string resources, both from
scratch (in the Resource Manager) and from Extracting Resource actions
in the editor for existing strings. Android strings require special
escaping in some circumstances (see
https://developer.android.com/guide/topics/resources/string-resource.html#escaping_quotes).
Existing code shows a mixture of escaped and unescaped strings in this
dialog, depending on the exact circumstance.

This change updates the logic so that the panel always displayed the
"plain text" version of a string, which is in line with what a user
would expect. The string is still later encoded appropriately before
it's written into the XML file.

Test: CreateXmlResourcePanelImplTest.java
Fixes: 196248641
Change-Id: Ib81a569b4b6094e9229a54bcffefe174ff208a6b