Memory reporting: limit shared components number
In theory the amount of shared component may be up to 2^NUMBER_OF_COMPONENTS. It's a huge number and in this case the size of the report will also be huge. Because of that we limit the number of shared components and will add only be biggest in the report. The value of the limit is stored in the server side flag in the shared_components_limit field

Test: None
Bug: None
Change-Id: Ife29794b502014f7f4d7ff3b2acd9d4fd0c91e38