Update model files now that AIDL is false by default
This change also modifies AndroidGradleTestUtils to allow tests to
specify whether to include AIDL source files (false by default).
This functionality is needed for the tests in
AidlDefaultRefactoringProcessorTest.

This change keeps the version in
AidlDefaultRefactoringProcessor.necessity() set to 8.0.0-alpha04
because that will be the version of AGP in which AIDL will be
disabled by default.

Bug: 243130937
Test: existing
Change-Id: I4d56b911622229df623d257e9c5a440a843a1496