[AUA] Notify the user of to-be-removed support for old AGPs
When considering compatiblity between Studio and AGP, there now exists
some possibly-empty region starting and during the
currently-compatible region (and consequently being met by the too-old
region).

For projects with AGP versions in this region, compute upgrade
requirements the same, except that the notification is more
strongly-worded.  (This means that the user may receive a suggestion
to upgrade to a deprecated version.)

(This is implemented with a new threshold version string in the common
Sdk, GRADLE_PLUGIN_NEXT_MINIMUM_VERSION.)

Bug: 180703127
Test: added, existing tests adjusted
Change-Id: I4533326e2c40133c7f90138838eeff7bdef66e20