Migrate DeviceExplorer to the DeviceProvisioner
AdbDeviceFileSystemService consumes the flow of DeviceHandles from the
Device Provisioner, wrapping them in AdbDeviceFileSystem objects, rather
than listening to AndroidDebugBridge.

AdbDeviceFileSystem is converted to hold a ConnectedDevice, obtained
from the Device Provisioner, rather than holding an IDevice and
obtaining a DeviceSelector at the lowest level.

DeviceHandleRenderer is used to render device names; this allows us to
remove the DeviceFileSystemRendererFactory.

Bug: 241591222
Bug: 205005362
Test: updated
Change-Id: Ib86299e8db8b65087ea1e3c082209f857d9f54cb