Flag resource change listener in Translations Editor
Unfortunately, this "fix" from earlier actually breaks the usage
of the Translations Editor. Because the updates from the editor
itself trigger reloads, the editor becomes unusable if you are
using the bar at the bottom. Put this behind a flag so that
we can continue to test it and the bits don't rot while we
figure out how to get this to work correctly.

Bug: 200817330
Test: None
Change-Id: Id4cf73cf069f1ee0763ee2382e9b7b65366240a8