[AUA] Enhance checks for namespace conflicts
Check the dsl- and manifest-provided values for namespace or
(particularly) testNamespace against values from the synced model, so
that a conflict detected through the model in the absence of Dsl
specification can also block this upgrade.

Also unconditionally emit testNamespace if a package directive is
removed from the androidTest manifest, even if the testNamespace is
the one that would be defaulted from the main namespace.  Because
testNamespace is now emitted even in the common case, adjust the
applicability of the processor, as testNamespace is not available in
the AGP 4.2.0 Dsl.

Bug: 191813691
Test: existing tests adjusted
Change-Id: Ie3079388c4f2dd55ac1c8c72385b62452f8fa2e8