[BA] remove saved data from BuildAttributionUiManager
This is a small step forward in simplifying this class.
Do not store BuildAttributionReportUiData anymore. It was needed
to re-open tab with the latest result after it was closed.
Now we have storage service responsible for this so get latest
data from there when necessary and change other places to pass
created ui data through parameters.
Note, we need to keep the boolean of if the latest build failed.
Storage does not have this information yet, it track only successful
builds ATM so this is a temporary tradeoff to keep this change
minimal and simple.

Bug: 235469262
Test: existing tests on BuildAttributionUiManager
Change-Id: I6fd269e4841ee873834cec96f6fe71b4bf3665fd