Tweak SDK source download text
Minor updates to the text displayed when SDK sources are missing, and
Android Studio offers to download them. To facilitate this, I'm changing
the code to only allow a single API level's sources to be downloaded at
a time via this mechanism (since that's all we're using in the code
anyway).

For reference, screenshot of the text in question is at: https://screenshot.googleplex.com/8LxbfQifWZ9V4cX

Bug: 74128300
Test: Included
Change-Id: Ied4c171c8bf432e391b321cf4e4819e5dffe38aa