Set the default JDK version to 17
as it is supposed to be. This version is used to find or create an entry
in the JDK table when the Android Studio Default JDK is not found.

Do it by setting it to the max supported version. Next time it changes
it may need to be set to an explicit version again and failing tests
should remind about it.

Bug: n/a
Test: n/a
Change-Id: Ic496eee697b5883fcd58f3bf56efefbbe6800d20